five  11/21

+++ src/main/java/com/wangdao/mall/controller/admin/AdminAdministerController.java
@@ -16,7 +16,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -43,7 +46,7 @@ public class AdminAdministerController {
      * @return
      */
     @RequestMapping("/admin/list")
-    @RequiresPermissions(value = {"admin:admin:list"})
+    @RequiresPermissions(value = {"admin:admin:list","admin:admin:update","damin:admin:read","admin:admin:delete","admin:admin:create"},logical = Logical.OR)
     public BaseReqVo selectAllAdmin(String username,Integer page,Integer limit,String sort,String order){
         BaseReqVo<Object> baseReqVo = new BaseReqVo<>();
         Map<String, Object> map = adminService.selectAllAdmin(username,page,limit,sort,order);
@@ -60,7 +63,7 @@ public class AdminAdministerController {
      * @return
      */
     @RequestMapping("/role/options")
-    @RequiresPermissions(value = {"admin:role/options"})
+    @RequiresPermissions(value = {"admin:admin:list","admin:admin:update","damin:admin:read","admin:admin:delete","admin:admin:create"},logical = Logical.OR)
     public BaseReqVo selectAllAdminRole(){
         List<Map> list = adminService.selectAllAdminRole();
         BaseReqVo<List<Map>> baseReqVo = new BaseReqVo<>(list, "成功", 0);
@@ -145,7 +148,7 @@ public class AdminAdministerController {
      * @return
      */
     @RequestMapping("/admin/update")
-    @RequiresPermissions(value = {"admin:admin:update"},logical = Logical.AND)
+    @RequiresPermissions(value = {"admin:admin:update"})
     public BaseReqVo updateAdmin(@RequestBody AdminDO adminDO){
         AdminDO adminDOAfterUpdate = adminService.updateAdmin(adminDO);
         if (adminDOAfterUpdate != null){
@@ -172,6 +175,12 @@ public class AdminAdministerController {
     @RequestMapping("/admin/delete")
     @RequiresPermissions(value = {"admin:admin:delete"})
     public BaseReqVo deleteAdmin(@RequestBody AdminDO adminDO){
+        HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        String adminName = (String) request.getSession().getAttribute("admin");
+        if (adminName.equals(adminDO.getUsername())){
+            BaseReqVo<Object> baseReqVoError = new BaseReqVo<>(null, "管理员不能删除自己帐号", 604);
+            return baseReqVoError;
+        }
         int result = adminService.deleteAdmin(adminDO);
         if (result != 0){
             BaseReqVo<Object> baseReqVo = new BaseReqVo<>(null, "成功", 0);
@@ -243,7 +252,7 @@ public class AdminAdministerController {
      * @return
      */
     @RequestMapping("/role/list")
-    @RequiresPermissions(value = {"admin:role:lsit"})
+    @RequiresPermissions(value = {"admin:role:list","admin:role:update","admin:role:read","admin:role:delete","admin:role:permission:get","admin:role:create","admin:role:permission:update"},logical = Logical.OR)
     public BaseReqVo selectAllRoleList(String name,Integer page,Integer limit,String sort,String order){
         Map<String,Object> map = adminService.selectAllRoleList(name,page,limit,sort,order);
         BaseReqVo<Map<String, Object>> baseReqVo = new BaseReqVo<>(map, "成功", 0);
@@ -394,7 +403,7 @@ public class AdminAdministerController {
      *  key 有问题
      */
     @RequestMapping("/storage/list")
-    @RequiresPermissions(value = {"admin:storage:list"})
+    @RequiresPermissions(value = {"admin:storage:list","admin:storage:update","admin:storage:read","admin:storage:delete","admin:storage:create"},logical = Logical.OR)
     public BaseReqVo selectAllStorageList(String name,String key,Integer page,Integer limit,String sort,String order){
         Map<String,Object> map = adminService.selectAllStorageList(name,key,page,limit,sort,order);
         BaseReqVo<Object> baseReqVo = new BaseReqVo<>(map, "成功", 0);
@@ -485,7 +494,7 @@ public class AdminAdministerController {
      * @return
      */
     @RequestMapping(value = "/role/permissions",method = {RequestMethod.GET})
-    @RequiresPermissions(value = {"admin:role:permissions"})
+    @RequiresPermissions(value = {"admin:role:permission:get","admin:role:permission:update"},logical = Logical.OR)
     public BaseReqVo listPermissionsByRoleId(Integer roleId){
         List<String> permissionList = adminService.listPermissions(roleId);
         List<SystemPermissionDO> systemPermissionDOList = adminService.systemPermissionsList();
@@ -498,7 +507,7 @@ public class AdminAdministerController {
 
 
     @RequestMapping(value = "/role/permissions",method = {RequestMethod.POST})
-    @RequiresPermissions(value = {"admin:role:permissions"})
+    @RequiresPermissions(value = {"admin:role:permission:update"})
     public BaseReqVo listPermissions(@RequestBody PermissionsVO permissionsVO){
         if (permissionsVO.getRoleId().equals(1)){
             BaseReqVo<Object> baseReqVo = new BaseReqVo<>(null, "你不能修改超级管理员的权限", 500);
